Refactor filename creation of images
